feature/ADAPT-3754 learner experience new file created
